pyproject.toml
[project]
name = "ageminibot"
version = "0.1.0"
description = "Telegram bot with Gemini grounding and image generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytelegrambotapi>=4.25.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "pymongo>=4.6.0",
    "huey>=2.5.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
]



dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.100.0",
    "mongomock>=4.1.0",
]
